idle/test_launcher.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Launcher for Settings API</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body/>

  <script type="application/javascript">
function getSiblingURI(myURI, siblingFilename, newDomain) {
  // Takes a URI and a filename, returns the URI of that filename if it were a sibling
  // of the original. Optionally replaces the domain as well.
  //
  // Ex: getSiblingURI("http://www.example.org/foo/bar.html", "fred.html")
  //       returns "http://www.example.org/foo/fred.html"

  // strip filename off URI
  resultURI = myURI.substr(0, myURI.lastIndexOf("/") + 1);

  // If domain specified, replace it
  if (newDomain) {
    resultURI = resultURI.replace(/:\/\/[^\/]*\//, "://" + newDomain + "/");
  }

  return resultURI + siblingFilename;
}
  var testURL = getSiblingURI(window.location.href, "idle_basics.html", "http://www.example.org");

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.addPermission("settings", true, testURL);

  var iframe = document.createElement("iframe");
  iframe.mozApp = "http://example.org/manifest.webapp";
  iframe.src = testURL;
  iframe.setAttribute("id", "ifrm");
  iframe.onload = loaded();
  document.body.appendChild(iframe);

  function loaded() {

    window.addEventListener("message", function onMessage(event) {
       if (event.data == "finished") {
         window.removeEventListener("message", onMessage, false);
         SimpleTest.finish();
       } else {
         if(event.data.type == "ok") {  
           ok(event.data.result, event.data.msg);
         }
         else if (event.data.type == "is") {
           is(event.data.result[0], event.data.result[1], event.data.msg);
         }
       }
           
     }, false);

  }
  </script>
</head>

</html>
